tidy: export methods to mixins
